refactor(text field): extract wrapper + action to input component
