+++ packages/react/src/components/TextField/TextField.styles.ts
@@ -6,17 +6,6 @@ import { TextFieldProps } from './TextField.props'
 
 type TextFieldStyleProps = Required<Pick<TextFieldProps, 'size' | 'feedback' | 'disabled' | 'readOnly' | 'action'>>
 
-const getFeedbackBorderColor = (theme: Theme, { feedback }: TextFieldStyleProps) => {
-  switch (feedback) {
-    case 'error':
-      return theme.color.alert
-    case 'success':
-      return theme.color.success
-    default:
-      return theme.color.lowEmphasis
-  }
-}
-
 const getFeedbackTextColor = (theme: Theme, { feedback, disabled }: TextFieldStyleProps) => {
   const defaultColor = disabled ? theme.color.lowEmphasis : theme.color.mediumEmphasis
 
@@ -30,45 +19,7 @@ const getFeedbackTextColor = (theme: Theme, { feedback, disabled }: TextFieldSty
   }
 }
 
-export const actionStyles = createUseStyles((theme: Theme) => ({
-  action: {
-    marginRight: ({ action }) => action === 'icon' && theme.spacing.tiny
-  },
-  actionImage: {
-    maxWidth: theme.size.large
-  }
-}))
-
-export const styles = createUseStyles((theme: Theme) => ({
-  inputContainer: {
-    position: 'relative',
-    backgroundColor: theme.color.surface,
-    border: '1px solid',
-    borderColor: (props: TextFieldStyleProps) => getFeedbackBorderColor(theme, props),
-    borderRadius: theme.borderRadius.medium,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    cursor: ({ disabled, readOnly }: TextFieldStyleProps) => (!disabled && !readOnly ? 'text' : 'default'),
-    '&:hover': {
-      borderColor: ({ disabled }: TextFieldStyleProps) => !disabled && theme.color.mediumEmphasis
-    },
-    '&:focus-within': {
-      border: '2px solid',
-      borderColor: theme.color.primary
-    },
-    '&:after': {
-      backgroundColor: theme.color.lowEmphasis,
-      content: '""',
-      height: '100%',
-      left: 0,
-      opacity: ({ readOnly }: TextFieldStyleProps) => (readOnly ? theme.opacity.disabledLow : 0),
-      position: 'absolute',
-      top: 0,
-      width: '100%',
-      pointerEvents: 'none'
-    }
-  },
+const styles = createUseStyles((theme: Theme) => ({
   label: {
     color: (props: TextFieldStyleProps) => getFeedbackTextColor(theme, props),
     display: 'inline-block',
@@ -91,3 +42,5 @@ export const styles = createUseStyles((theme: Theme) => ({
     }
   }
 }))
+
+export default styles
